backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
apscheduler==3.11.0
requests==2.32.5
selenium==4.35.0
webdriver-manager==4.0.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
